Configuracion y orden de entity

+++ Datos/Models/Bitacora.cs
@@ -2,7 +2,7 @@
 {
     public class Bitacora
     {
-        public int idBitacora { get; set; }
+        public long idBitacora { get; set; }
         public string jsonEntrada { get; set; }
         public string Estado { get; set; }
         public long idConversation { get; set; }
